AI Workflow Functions/fetch_news_data/requirements.txt
functions-framework
requests
beautifulsoup4
cloud-sql-python-connector[pg8000]
sqlalchemy
pg8000
pandas
google-cloud-secret-manager
FinNews